Add WaitBB status (#10)
* Add a test for issue#138

* Make new players with status `Waitbb`

Newly joined players are now initialized with status Waitbb,
instead of `Init`, except in the situation where the new players
are the first and second to join.

Players with status `Waitbb` cannot play the game and must wait
until their alloted positions can be the big blind of next hand.

* Update

* Make WaitBB player the actual BB in 3 situations

* Add tests for `arrange_waitbbs`

* Clean useless logs

---------

Co-authored-by: Linerre <leon.errelin@outlook.com>